move maybe-usable-by-other-cc-components to pkg. Fix all files to use the new paths
