+++ pkg/multiChanTicker/README.md
