+++ receivers/receiveManager.go
@@ -5,8 +5,8 @@ import (
 	"os"
 	"sync"
 
-	cclog "github.com/ClusterCockpit/cc-metric-collector/internal/ccLogger"
-	lp "github.com/ClusterCockpit/cc-metric-collector/internal/ccMetric"
+	cclog "github.com/ClusterCockpit/cc-metric-collector/pkg/ccLogger"
+	lp "github.com/ClusterCockpit/cc-metric-collector/pkg/ccMetric"
 )
 
 var AvailableReceivers = map[string]func(name string, config json.RawMessage) (Receiver, error){
