+++ sinks/gangliaSink.go
@@ -9,8 +9,8 @@ import (
 	//	"time"
 	"os/exec"
 
-	cclog "github.com/ClusterCockpit/cc-metric-collector/internal/ccLogger"
-	lp "github.com/ClusterCockpit/cc-metric-collector/internal/ccMetric"
+	cclog "github.com/ClusterCockpit/cc-metric-collector/pkg/ccLogger"
+	lp "github.com/ClusterCockpit/cc-metric-collector/pkg/ccMetric"
 )
 
 const GMETRIC_EXEC = `gmetric`
