+++ sinks/stdoutSink.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	//	"time"
-	lp "github.com/ClusterCockpit/cc-metric-collector/internal/ccMetric"
+	lp "github.com/ClusterCockpit/cc-metric-collector/pkg/ccMetric"
 )
 
 type StdoutSink struct {
